app/requirements.txt
flask >= 2.2.2
python-dotenv